only admin can see link to modify && delete in :arrow_right: layouts/news + add comment in :arrow_right: deleted_tasks.php

+++ layouts/news.php
@@ -23,6 +23,16 @@ include('../Database/connection_database.php');
     <?php
         include('../components/header.php');
         include('../components/navBar_categories.php');
+        
+        /* -------------------- VERIFY IS USER IS ADMIN --------------------*/
+
+        $verify_user_is_admin = $bdd->prepare('SELECT is_admin FROM user WHERE id =:id_user');
+        $verify_user_is_admin->bindParam(':id_user', $_SESSION['id_user']);
+        $verify_user_is_admin->execute();
+
+        $user_admin = $verify_user_is_admin->fetch();
+
+        /* -------------------- DISPLAY NEWS --------------------*/
 
         $select_articles = $bdd->prepare('SELECT * FROM whatsnew');
         $select_articles->execute();
@@ -32,22 +42,30 @@ include('../Database/connection_database.php');
         <?php
         while($display_articles = $select_articles->fetch()){
             ?>
-                    <h1 class="underlineH1">
-                        <?php 
-                            echo $display_articles['currentdate'];
-                        ?>
-                    </h1>
+                <h1 class="underlineH1">
+                    <?php 
+                        echo $display_articles['currentdate'];
+                    ?>
+                </h1>
 
-                    <p>
-                        <?php
-                            echo $display_articles['news'];
-                        ?>
-                    </p>
-
-                    <p>
-                        <a href="./Admin/news_modify.php?id_news=<?php echo $display_articles['id']; ?>">Modify</a> |
-                        <a href="../Database/Admin/deleteNews_database.php?id=<?php echo $display_articles['id']; ?>">Delete</a>
-                    </p>
+                <p>
+                    <?php
+                        echo $display_articles['news'];
+                    ?>
+                </p>
+        
+                <?php
+                    /* -------------------- IF USER IS ADMIN ADD LINK FOR MODIFY && DELETE --------------------*/
+                    
+                    if($user_admin['is_admin']){
+                    ?>
+                        <p>
+                            <a href="./Admin/news_modify.php?id_news=<?php echo $display_articles['id']; ?>">Modify</a> |
+                            <a href="../Database/Admin/deleteNews_database.php?id=<?php echo $display_articles['id']; ?>">Delete</a>
+                        </p>
+                    <?php
+                    }
+                ?>
             <?php
         }
     ?>
